Tiny change to BoardListener, I forgot to cast the coordinates of the Gizmo to an (int) when we were renaming a Gizmo after it had moved board position. Also added a file that we could probably use to easily and quickly demonstrate the 3 bonus gizmos to Kostas.

+++ src/controller/BoardListener.java
@@ -151,7 +151,7 @@ public class BoardListener implements MouseListener {
 
 						pm.getGizmoToMove().moveGizmo(x, y);
 						// Added to update a Gizmos name when it is moved to a new position on the board
-						pm.getGizmoToMove().setName(""+pm.getGizmoToMove().getXPos()+"_"+pm.getGizmoToMove().getYPos());
+						pm.getGizmoToMove().setName(""+(int)pm.getGizmoToMove().getXPos()+"_"+(int)pm.getGizmoToMove().getYPos());
 						
 
 						int numberOfRotations = pm.getGizmoToMove().getGizAngle() / 90;
